# individuals-api/openapi/individuals-api.yaml
openapi: 3.0.3

info:
  title: individuals-api
  description: |
    External entry layer of the payment platform.

    Orchestrates registration and authentication. Owns no domain data:
      - person-service is the source of truth about the domain user
      - Keycloak is the source of truth about the account, tokens and roles
  version: 1.0.0
  contact:
    name: dezxxx
    email: web7tudio@gmail.com
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0

servers:
  - url: http://localhost:8081
    description: Local

tags:
  - name: auth
    description: Registration and authentication

paths:

  /v1/auth/registration:
    post:
      tags: [auth]
      operationId: register
      summary: Register a new user
      description: |
        The order is fixed. person-service creates the domain user and issues
        its user_uid - the domain identifier is issued by the domain service,
        never here. The Keycloak account is created next, carrying that
        user_uid as an attribute; its password is set by a second Admin call,
        because the create call does not accept one. The new user is then
        logged in, so no separate login is needed.

        The two systems share no transaction. An account created without a
        password is deleted again; a domain user that outlives a failed
        Keycloak account cannot be, and is reported as REGISTRATION_INCONSISTENT.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationRequest"
            examples:
              valid:
                summary: Well-formed registration
                value:
                  email: user@dezxxx.com
                  password: Str0ngP@ssw0rd
                  confirmPassword: Str0ngP@ssw0rd
                  firstName: Ivan
                  lastName: Ivanov
              mismatchedConfirmation:
                summary: Rejected with 400 - confirmation does not match
                value:
                  email: user@dezxxx.com
                  password: Str0ngP@ssw0rd
                  confirmPassword: Str0ngP@ssw0rt
                  firstName: Ivan
                  lastName: Ivanov
      responses:
        "201":
          description: User registered and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
              examples:
                registered:
                  summary: Tokens issued right after registration
                  value:
                    accessToken: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9
                    refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
                    expiresIn: 300
                    tokenType: Bearer
                    userUid: 6f1d2a4e-8c3b-4a7f-9e2d-1b5c7a9f0e33
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  /v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Log in with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              valid:
                summary: Known user
                value:
                  email: user@dezxxx.com
                  password: Str0ngP@ssw0rd
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
              examples:
                authenticated:
                  value:
                    accessToken: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9
                    refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
                    expiresIn: 300
                    tokenType: Bearer
                    userUid: 6f1d2a4e-8c3b-4a7f-9e2d-1b5c7a9f0e33
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  /v1/auth/refresh-token:
    post:
      tags: [auth]
      operationId: refreshToken
      summary: Exchange a refresh token for a new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
            examples:
              valid:
                summary: Refresh token issued by a previous login
                value:
                  refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
              examples:
                refreshed:
                  value:
                    accessToken: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9
                    refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
                    expiresIn: 300
                    tokenType: Bearer
                    userUid: 6f1d2a4e-8c3b-4a7f-9e2d-1b5c7a9f0e33
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  /v1/auth/me:
    get:
      tags: [auth]
      operationId: getCurrentUser
      summary: Get the currently authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserResponse"
              examples:
                current:
                  summary: Resolved from the bearer token
                  value:
                    userUid: 6f1d2a4e-8c3b-4a7f-9e2d-1b5c7a9f0e33
                    keycloakUserId: 0b7a5f2c-19d4-4a8e-9c3b-77e1a0d4f5b6
                    email: user@dezxxx.com
                    firstName: Ivan
                    lastName: Ivanov
                    emailVerified: false
                    roles: [USER]
                    registeredAt: "2026-08-19T12:34:56Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    RegistrationRequest:
      type: object
      required: [email, password, confirmPassword, firstName, lastName]
      # OpenAPI has no way to say "this field equals that one", so the
      # confirmation rule is the single constraint of this schema that the
      # keywords below cannot express. The extension stamps our class-level
      # constraint onto the generated model, which keeps the rule in the
      # contract rather than in a controller someone has to remember to call.
      # Fully qualified on purpose: the generator copies the line verbatim and
      # adds no import for it.
      x-class-extra-annotation: "@com.dezxxx.individuals.validation.PasswordsMatch"
      properties:
        email:
          type: string
          format: email
          maxLength: 1024
          example: user@dezxxx.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          example: Str0ngP@ssw0rd
        confirmPassword:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          example: Str0ngP@ssw0rd
        # Required because person.users declares first_name and last_name
        # NOT NULL - the domain user cannot be created without them.
        firstName:
          type: string
          minLength: 1
          maxLength: 64
          example: Ivan
        lastName:
          type: string
          minLength: 1
          maxLength: 64
          example: Ivanov

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 1024
          example: user@dezxxx.com
        # No minLength here on purpose. Password policy is enforced at
        # registration and by Keycloak; rejecting a short attempt with 400
        # instead of 401 would leak the policy and would lock out existing
        # users if it ever loosened. maxLength stays only as a payload guard.
        password:
          type: string
          format: password
          maxLength: 72
          example: Str0ngP@ssw0rd

    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 1
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9

    TokenResponse:
      type: object
      required: [accessToken, refreshToken, expiresIn, tokenType, userUid]
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
        expiresIn:
          type: integer
          format: int64
          description: Access token lifetime in seconds
          example: 300
        tokenType:
          type: string
          example: Bearer
        userUid:
          type: string
          format: uuid
          description: Domain user identifier issued by person-service
          example: 6f1d2a4e-8c3b-4a7f-9e2d-1b5c7a9f0e33

    CurrentUserResponse:
      type: object
      # All of these are always known for an authenticated user: the token
      # carries sub and the roles, and the domain columns behind firstName and
      # lastName are NOT NULL. Matches the payload the handout shows for /me.
      required: [userUid, keycloakUserId, email, firstName, lastName, emailVerified, roles, registeredAt]
      properties:
        userUid:
          type: string
          format: uuid
          description: The platform identifier - the only one the business layer uses
          example: 6f1d2a4e-8c3b-4a7f-9e2d-1b5c7a9f0e33
        keycloakUserId:
          type: string
          maxLength: 64
          description: |
            Keycloak `sub`. A technical link to the account, never the
            platform's primary identifier.
          example: 0b7a5f2c-19d4-4a8e-9c3b-77e1a0d4f5b6
        email:
          type: string
          format: email
          example: user@dezxxx.com
        # person-service owns the domain profile; the account carries a copy
        # written at registration, and this answer is built from that copy
        # because module 1 has no read side to ask. So the two names can lag
        # behind a profile edited elsewhere, and the copy is not the truth.
        firstName:
          type: string
          maxLength: 64
          description: As recorded on the account. May lag the domain profile.
          example: Ivan
        lastName:
          type: string
          maxLength: 64
          description: As recorded on the account. May lag the domain profile.
          example: Ivanov
        emailVerified:
          type: boolean
          example: false
        roles:
          type: array
          items:
            type: string
          example: [USER]
        registeredAt:
          type: string
          format: date-time
          description: |
            When the account was created in Keycloak. Not present in the access
            token, which is why /me reads the profile through the Admin API
            instead of answering from the token alone.
          example: "2026-08-19T12:34:56Z"

    # Single error model shared by every service of the platform.
    ErrorResponse:
      type: object
      required: [timestamp, path, status, error, message, traceId]
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-08-19T12:34:56Z"
        path:
          type: string
          example: /v1/auth/registration
        status:
          type: integer
          format: int32
          example: 400
        error:
          type: string
          description: Stable machine-readable error code
          example: VALIDATION_ERROR
        message:
          type: string
          example: Password confirmation does not match
        traceId:
          type: string
          description: Correlates the response with the trace stored in Tempo
          example: 4bf92f3577b34da6a3ce929d0e0e4736
        details:
          type: array
          description: Field-level failures, one readable line each
          items:
            type: string
          example:
            - "confirmPassword: must match password"

  responses:
    BadRequest:
      description: Request failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validationFailed:
              value:
                timestamp: "2026-08-19T12:34:56Z"
                path: /v1/auth/registration
                status: 400
                error: VALIDATION_ERROR
                message: Password confirmation does not match
                traceId: 4bf92f3577b34da6a3ce929d0e0e4736
                details:
                  - "confirmPassword: must match password"
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            badCredentials:
              value:
                timestamp: "2026-08-19T12:34:56Z"
                path: /v1/auth/login
                status: 401
                error: INVALID_CREDENTIALS
                message: Email or password is incorrect
                traceId: 4bf92f3577b34da6a3ce929d0e0e4736
    Conflict:
      description: User already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            duplicate:
              value:
                timestamp: "2026-08-19T12:34:56Z"
                path: /v1/auth/registration
                status: 409
                error: USER_ALREADY_EXISTS
                message: Email is already registered
                traceId: 4bf92f3577b34da6a3ce929d0e0e4736
    NotFound:
      description: |
        The token is valid, but the account it points at no longer exists in
        Keycloak - deleted, or moved to another realm.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missing:
              value:
                timestamp: "2026-08-19T12:34:56Z"
                path: /v1/auth/me
                status: 404
                error: NOT_FOUND
                message: Resource not found
                traceId: 4bf92f3577b34da6a3ce929d0e0e4736
    InternalServerError:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            unexpected:
              summary: Nothing leaks outwards but the trace id
              value:
                timestamp: "2026-08-19T12:34:56Z"
                path: /v1/auth/registration
                status: 500
                error: INTERNAL_ERROR
                message: Unexpected error
                traceId: 4bf92f3577b34da6a3ce929d0e0e4736
    ServiceUnavailable:
      description: A downstream dependency is unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            keycloakDown:
              value:
                timestamp: "2026-08-19T12:34:56Z"
                path: /v1/auth/registration
                status: 503
                error: DEPENDENCY_UNAVAILABLE
                message: Keycloak is not reachable
                traceId: 4bf92f3577b34da6a3ce929d0e0e4736